#3082 PR (Please Test First)
I could not, after hours, get Docker to cooperate and let me access the website. So with how simple this change is, I, after much deliberation and exasperation, have decided that it probably works. But on the off chance it doesn't can somebody please run this and make sure it actually works as intended?